% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hdr_nii.R
\name{convert_hdr_nii}
\alias{convert_hdr_nii}
\title{Convert the data from a Header \code{data.frame} into NIfTI files}
\usage{
convert_hdr_nii(sub_hdr, merge_files = TRUE, ignore_derived = TRUE,
  overwrite = FALSE, rename = TRUE, opts = "-f \%t")
}
\arguments{
\item{sub_hdr}{Output from \code{\link{subset_hdr}} or the
data relevant from \code{\link{noncon_brain_ct}}}

\item{merge_files}{Should files be merged, passed do \code{dcm2nii} options}

\item{ignore_derived}{Should derived images be ignored,
passed do \code{dcm2nii} options}

\item{overwrite}{Should files be overwritten if they exist}

\item{rename}{Should the files be renamed}
}
\value{
A list of the output filenames and the result for each directory
\code{dcm2nii} call
}
\description{
Convert the data from a Header \code{data.frame} into NIfTI files
}
