% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcmsort.R
\name{dcmsort}
\alias{dcmsort}
\title{Sort DICOM directories}
\usage{
dcmsort(directory, copy_files = FALSE, remove_original = TRUE,
  remove_directories = TRUE, remove_compressed = TRUE, ...)
}
\arguments{
\item{directory}{Directory of DICOM files/folders}

\item{copy_files}{Should the files be copied to a temporary location before
running.  Alternatively, you can set this to \code{FALSE} and
do this beforehand.  If \code{FALSE}, the data will be deleted from
original directory!}

\item{remove_original}{Should original files be deleted? This is passed to
\code{\link{sortDicomDirectories}}}

\item{remove_directories}{Should all the old directories be removed after
sorting}

\item{remove_compressed}{Remove folders that end in
\code{zip|rar|gz|tar|bz|xz}}

\item{...}{additional arguments to \code{\link{sortDicomDirectories}}}
}
\value{
A List of the directory and the old/new directories
}
\description{
Sort DICOM directories
}
