% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_tags.R
\name{relevant_tags}
\alias{relevant_tags}
\alias{relevant_fields}
\title{Tags to Keep for Filtering}
\usage{
relevant_tags()

relevant_fields()
}
\value{
Character vector
}
\description{
Tags to Keep for Filtering
}
\examples{
relevant_tags()
}
