% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_hdr.R
\name{subset_hdr}
\alias{subset_hdr}
\title{Subset the Header files and Reshape}
\usage{
subset_hdr(all_hdr, keep_tags = relevant_tags(),
  unique_fields = relevant_fields(), force = TRUE)
}
\arguments{
\item{all_hdr}{An output from \code{\link{read_all_hdr}}}

\item{keep_tags}{tags to keep in the data}

\item{unique_fields}{The fields that should uniquely identify a DICOM}

\item{force}{Should the function warn (as opposed to fail)
if \code{unique_fields} has
duplicates?}
}
\value{
A wide \code{data.frame} of the information needed for filtering
}
\description{
Subset the Header files and Reshape
}
