% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_hdr.R
\name{read_all_hdr}
\alias{read_all_hdr}
\title{Read all DICOM headers in a directory and return a \code{data.frame}}
\usage{
read_all_hdr(directory, outfile = tempfile(fileext = ".rds"),
  overwrite = FALSE)
}
\arguments{
\item{directory}{Directory of DICOM files}

\item{outfile}{Output RDS file to write.  If it exists and
\code{overwrite = FALSE}, then it will be read in and returned.}

\item{overwrite}{Should the RDS file be overwritten if it exists}
}
\value{
A \code{data.frame} of all the tags
}
\description{
Read all DICOM headers in a directory and return a \code{data.frame}
}
\seealso{
\code{\link{read_dicom_header}}
}
